+++ rust/candid/src/lib.rs
@@ -246,6 +246,8 @@
 //! ```
 //!
 
+#![allow(clippy::upper_case_acronyms)]
+
 pub use candid_derive::{candid_method, export_service, CandidType};
 pub use serde::Deserialize;
 
@@ -262,6 +264,7 @@ pub use types::CandidType;
 pub use types::{
     number::{Int, Nat},
     principal::Principal,
+    reference::{Func, Service},
     reserved::{Empty, Reserved},
 };