+++ rust/candid/tests/assets/ok/fieldnat.d.ts
@@ -1,12 +1,15 @@
 import type { Principal } from '@dfinity/agent';
-import type BigNumber from 'bignumber.js';
+export interface non_tuple {
+  _1_ : string,
+  _2_ : string,
+};
+export type tuple = [string, string];
 export default interface _SERVICE {
-  'bab' : (arg_0: BigNumber, arg_1: BigNumber) => Promise<undefined>,
-  'bar' : (arg_0: { '2' : BigNumber }) => Promise<undefined>,
-  'bas' : (arg_0: [BigNumber, BigNumber]) => Promise<[string, BigNumber]>,
-  'baz' : (arg_0: { _2_ : BigNumber, '2' : BigNumber }) => Promise<{}>,
-  'bib' : (arg_0: [BigNumber]) => Promise<{ _0_ : BigNumber }>,
-  'foo' : (arg_0: { _2_ : BigNumber }) => Promise<
-      { _2_ : BigNumber, '_2' : BigNumber }
-    >,
+  'bab' : (arg_0: bigint, arg_1: bigint) => Promise<undefined>,
+  'bar' : (arg_0: { '2' : bigint }) => Promise<undefined>,
+  'bas' : (arg_0: [bigint, bigint]) => Promise<[string, bigint]>,
+  'baz' : (arg_0: { _2_ : bigint, '2' : bigint }) => Promise<{}>,
+  'bba' : (arg_0: tuple) => Promise<non_tuple>,
+  'bib' : (arg_0: [bigint]) => Promise<{ _0_ : bigint }>,
+  'foo' : (arg_0: { _2_ : bigint }) => Promise<{ _2_ : bigint, '_2' : bigint }>,
 };