+++ rust/candid/tests/assets/ok/recursion.d.ts
@@ -1,19 +1,16 @@
 import type { Principal } from '@dfinity/agent';
-import type BigNumber from 'bignumber.js';
 export type A = B;
 export type B = [] | [A];
 export type list = [] | [node];
-export interface node { 'head' : BigNumber, 'tail' : list };
+export interface node { 'head' : bigint, 'tail' : list };
 export interface s {
   'f' : t,
   'g' : (arg_0: list) => Promise<[B, tree, stream]>,
 };
-export type stream = [] | [
-  { 'head' : BigNumber, 'next' : [Principal, string] }
-];
+export type stream = [] | [{ 'head' : bigint, 'next' : [Principal, string] }];
 export type t = (arg_0: Principal) => Promise<undefined>;
 export type tree = {
-    'branch' : { 'val' : BigNumber, 'left' : tree, 'right' : tree }
+    'branch' : { 'val' : bigint, 'left' : tree, 'right' : tree }
   } |
-  { 'leaf' : BigNumber };
+  { 'leaf' : bigint };
 export default s;